kos/image_builder/env/src/env.c: append env with NODE_TEST_DIR value
So far the only 'client' for the 'env' module is the nodejs tests. In
order to have proper variable to be passed to test.py we need to specify
it over env.c. Just to not rebuild KOS-image each time, as we would if we
put these variables to init.yaml (so in case if we run tests oer test.py
the init.yaml is ignored). So eventually this saves test running time
significantly.

Signed-off-by: Vadim Lomovtsev <vadim.lomovtsev@auriga.com>