kos/rules/make-images.mk: put nodejs executable dummy to the ramdisk
Some tests from the NodeJS test suite checks(expects) that there are
nodejs (node) binary is available on the filesystem, while for KOS
ideology, the executables are placed to the romfs image that is not
explicitly mounted to the user application space.

In order to pass such tests, we need to place fake executable (with the
same name as the real one) on the ramdisk image, so that tests
successfully checks it's existence.

Signed-off-by: Vadim Lomovtsev <vadim.lomovtsev@auriga.com>