test/common/tmpdir.js: fix path issue
While running test suite (or single test) the test system of nodejs can
use externally provided path to temporary directory. The actual path is
provided over NODE_TEST_DIR (it is) environment variable.

However, it was found that the tmpdir.js has an issue with evaluating
eventual path with then leads to removing '/' symbol from the very
beginning of the path, so then some tests fails.

This commit is to update evaluation of the test temporary directory
name with the one provided over NODE_TEST_DIR variable.

Signed-off-by: Vadim Lomovtsev <vadim.lomovtsev@auriga.com>