src: let the NODE_IMPLEMENTS_POSIX_CREDENTIALS defined for KOS-build
Some credentials API is not available at POSIX, at the KasperskyOS
and for other OSes. So the NodeJS has it's own abstraction for that,
which is guarded by the NODE_IMPLEMENTS_POSIX_CREDENTIALS.

This commit is to allow this macro to be used for KOS-build. However
it is required to disable some subsequent calls to initgrous() and
setgroups() functions for such implementation.

Signed-off-by: Vadim Lomovtsev <vadim.lomovtsev@auriga.com>